+++ tests/unit_tests/test_record_id.py
@@ -11,8 +11,7 @@ class TestRecordId(TestCase):
         self.assertEqual(record_id.id_part, "tobie")
         self.assertEqual(record_id.table_part, "person")
         self.assertEqual(record_id.to_prefixed_string(), "r'person:tobie'")
-        self.assertEqual(record_id.to_uid_string(), "person:⟨tobie⟩")
-        self.assertEqual(record_id.to_uid_string_with_backticks(), "person:`tobie`")
+        self.assertEqual(record_id.to_uid_string(), "person:`tobie`")
         self.assertEqual(str(record_id), "RecordId('person:tobie')")
 
     def test_create_with_table(self):
@@ -21,7 +20,7 @@ class TestRecordId(TestCase):
         self.assertEqual(record_id.id_part, "tobie")
         self.assertEqual(record_id.table_part, "person")
         self.assertEqual(record_id.to_prefixed_string(), "r'person:tobie'")
-        self.assertEqual(record_id.to_uid_string(), "person:⟨tobie⟩")
+        self.assertEqual(record_id.to_uid_string(), "person:`tobie`")
         self.assertEqual(str(record_id), "RecordId('person:tobie')")
 
     def test_raise_no_colon(self):
@@ -42,7 +41,7 @@ class TestRecordId(TestCase):
         record_id = RecordId('person:100')
         self.assertEqual(record_id.to_valid_string(), "person:100")
         record_id = RecordId('person:8424486b-85b3-4448-ac8d-5d51083391c7')
-        self.assertEqual(record_id.to_valid_string(), "person:⟨8424486b-85b3-4448-ac8d-5d51083391c7⟩")
+        self.assertEqual(record_id.to_valid_string(), "person:`8424486b-85b3-4448-ac8d-5d51083391c7`")
 
 
 if __name__ == '__main__':